pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chat-agent-api"
version = "0.1.0"
description = "A FastAPI chat agent backend powered by Pydantic AI"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Ashleigh Hewitt" }
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pydantic-ai[anthropic]>=0.0.12",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "loguru>=0.7.0",
    "redis>=5.0.0",
    "opentelemetry-api>=1.0.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-sdk>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "alembic>=1.13.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "alembic>=1.13.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
